+++ src/mount.ts
@@ -10,14 +10,14 @@ interface MountableFunctionOptions {
  * Mounts the mountable function on the elements.
  */
 export function mount (
-	selector: string|HTMLElement[],
+	selector: string|HTMLElement|HTMLElement[],
 	mountable: MountableFunction,
 	options: MountableFunctionOptions = {}
 ) : void
 {
 	const elements = typeof selector === "string"
 		? find(selector, options.context || document)
-		: selector;
+		: Array.isArray(selector) ? selector : [selector];
 
 	elements.forEach(element => mountable(element, ...(options.args || [])));
 }