+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 *   (feature) Add `safeParseJson()`.
 *   (feature) Add `parseElementContentAsJson()`.
+*   (improvement) Allow to pass a single element to `mount()`.
 
 
 1.2.1