+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+1.3.0
+=====
+
+*   (feature) Add `safeParseJson()`.
+*   (feature) Add `parseElementContentAsJson()`.
+*   (improvement) Allow to pass a single element to `mount()`.
+*   (feature) Add `mountJsx()`.
+
+
 1.2.1
 =====